server_ai/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
python-multipart
httpx
python-dotenv
psycopg2-binary
